requirements.txt
flask
certifi
google-genai
ollama
openai
pillow
python-docx
